move mobile_kpi_support_metrics generator logic into separate module
